+++ example/lib/main.dart
@@ -18,6 +18,7 @@ import 'inapp.dart';
 import 'second_page.dart';
 import 'utils.dart';
 
+// ignore_for_file: deprecated_member_use
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   // FirebaseApp not configured for web app. Added the check to avoid run time errors.
@@ -668,6 +669,24 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
                   onTap: () {
                     _moengagePlugin.registerForProvisionalPush();
                   }),
+              ListTile(
+                  title: const Text('Identify User (String)'),
+                  onTap: () {
+                    _moengagePlugin.identifyUser("flutter-uid");
+                  }),
+              ListTile(
+                  title: const Text('Identify User (Map)'),
+                  onTap: () {
+                    _moengagePlugin.identifyUser(
+                        {"email": "flutter@moengage.com", "id": "flutter"});
+                  }),
+              ListTile(
+                  title: const Text('Get Identities'),
+                  onTap: () async {
+                    Map<String, String>? identities =
+                        await _moengagePlugin.getUserIdentities();
+                    debugPrint('$tag Main : User Identities $identities');
+                  })
             ]).toList(),
           ),
         ),