+++ packages/moengage_flutter/moengage_flutter_ios/CHANGELOG.md
@@ -1,4 +1,9 @@
 # MoEngage Flutter iOS Plugin
+
+# Release Date
+
+## Release Version
+- Adding support for identify user based on multiple identities
           
 # 25-11-2024